Have filter functioning semi
