+++ js/app.js
@@ -2,8 +2,11 @@
 /* jslint esnext: true */
 var angular = require('angular');
 var angularRoute = require('angular-route');
+require('./service');
 require('./filters');
-var app = angular.module('NewsFeedApp', ['ngRoute', 'publisherName']);
+
+
+var app = angular.module('NewsFeedApp', ['ngRoute', 'publisherName', 'NewsService']);
 
 app.filter('reverse', function() {
   return function(items) {
@@ -65,64 +68,48 @@ app.controller('SavedViewController', ['$scope', '$http', 'NewsService', functio
 }]);
 
 /*Create a service to store the data*/
-app.factory('NewsService', function ($http) {
-
-    let stories = [];
-    let savedStories = [];
-    let publishers = [];
-    let interests = [];
-
-    /*Any data we need to store can go in here.*/
-    $http({
-            method: 'get',
-            url: 'http://chat.queencityiron.com/api/news/latest',
-        })
-        .then(function (response) {
-            console.log('Response of stories');
-            console.log(response);
-            angular.copy(response.data.stories, stories);
-        });
-
-    /*Testing to see if the publisher's data is cached to improve performance.*/
-    //    if (publishers.length === 0) {
-    //        $http({
-    //                method: 'get',
-    //                url: 'http://chat.queencityiron.com/api/publishers',
-    //            })
-    //            .then(function (response) {
-    //                publishers.push(response.data.providers);
-    //            })
-    //            .then(function () {
-    //                console.log('Publisher information saved');
-    //                console.log(publishers);
-    //            });
-    //    }
-
-    return {
-        getArticles: function () {
-            return stories;
-        },
-        saveArticle: function (article) {
-            /*Write code to push the saved article to an array.*/
-            savedStories.push(article);
-            console.log('article saved.');
-        },
-        getSavedArticles: function () {
-            return savedStories;
-        },
-        /*Not yet calling this function.*/
-        getPublishers: function () {
-            console.log('GetPublishers');
-            return publishers;
-        },
-        saveInterest: function (interest) {
-          console.log('Interest Saved!');
-          interests.push(interest);
-          console.log(interests);
-        },
-        getInterests: function () {
-          return interests;
-        },
-    };
-
-});
+//app.factory('NewsService', function ($http) {
+//
+//    let stories = [];
+//    let savedStories = [];
+//    let publishers = [];
+//    let interests = [];
+//
+//    /*Any data we need to store can go in here.*/
+//    $http({
+//            method: 'get',
+//            url: 'http://chat.queencityiron.com/api/news/latest',
+//        })
+//        .then(function (response) {
+//            console.log('Response of stories');
+//            console.log(response);
+//            angular.copy(response.data.stories, stories);
+//        });
+//    return {
+//        getArticles: function () {
+//            return stories;
+//        },
+//        saveArticle: function (article) {
+//            /*Write code to push the saved article to an array.*/
+//            savedStories.push(article);
+//            console.log('article saved.');
+//        },
+//        getSavedArticles: function () {
+//            return savedStories;
+//        },
+//        /*Not yet calling this function.*/
+//        getPublishers: function () {
+//            console.log('GetPublishers');
+//            return publishers;
+//        },
+//        saveInterest: function (interest) {
+//          console.log('Interest Saved!');
+//          interests.push(interest);
+//          console.log(interests);
+//        },
+//        getInterests: function () {
+//          return interests;
+//        },
+//    };
+//
+//});
